Punkt 12 - publiczne klasy
